requirements.txt
requests==2.28.2
Flask==2.2.1
Flask-SQLAlchemy==3.0.3
PyYAML==6.0
ua-parser==0.16.1
user-agents==2.2.0
py-mmdb-encoder==1.0.4
maxminddb==2.2.0
netaddr==0.8.0
## The following requirements were added by pip freeze:
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
SQLAlchemy==2.0.3
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
mysql-connector